feat(refactor): Extract line position calculation to helper and implement ts-morph text extraction
